fix: missing outros in kind of service array
